requirements.txt
firecrawl-py>=1.0
pydantic>=2.0
click>=8.0
python-dotenv>=1.0
httpx>=0.27
fastapi>=0.110
uvicorn[standard]>=0.29
apscheduler>=3.10
